requirements.txt
Fabric==1.8.0
Flask==0.10.1
Flask-Sockets==0.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
beautifulsoup4==4.3.1
ecdsa==0.9
gevent==0.13.8
gevent-websocket==0.3.6
greenlet==0.4.1
itsdangerous==0.23
paramiko==1.12.0
pycrypto==2.6
pymongo==2.6.2
python-dateutil==2.1
pytz==2013.6
requests==2.0.0
six==1.4.1
wsgiref==0.1.2
